requirements.txt
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
python-dateutil>=2.8.2
PyGithub>=2.1.1
Pillow>=10.0.0
svgwrite>=1.4.3
pytz>=2023.3